fix: checkbox item, style
