+++ rslang/src/app/ebook/components/ebook-settings/ebook-settings.component.html
@@ -1,32 +1,32 @@
 <section class="ebook-settings">
-    <form #settingsForm="ngForm">
-        <div *ngIf="isUserAuthenticated">
-            <mat-form-field class="example-full-width">
-                <input matInput id="username" value="User name" />
-            </mat-form-field>
-        </div>
-        <p class=" subtitle ">Buttons</p>
-        <div *ngFor="let itembtn of userBookSettings.buttonOptions ">
-            <input type="checkbox " id="itembtn.value " name="itembtn.value " [(ngModel)]="itembtn.checked " />
-            <label for="itembtn.value ">
-              {{ itembtn.label }}
-              <svg viewBox="0 0 100 100 " xmlns="http://www.w3.org/2000/svg ">
-                <path d="M 10 10 L 90 90 "></path>
-                <path d="M 90 10 L 10 90 "></path>
-              </svg>
-            </label>
-        </div>
-        <p class="subtitle ">Info about word</p>
-        <div *ngFor="let itemword of userBookSettings.wordOptions ">
-            <input type="checkbox " id="itemword.value " name="itemword.value " [(ngModel)]="itemword.checked " />
-            <label for="itemword.value ">
-              {{ itemword.label }}
-              <svg viewBox="0 0 100 100 " xmlns="http://www.w3.org/2000/svg ">
-                <path d="M 10 10 L 90 90 "></path>
-                <path d="M 90 10 L 10 90 "></path>
-              </svg>
-            </label>
-        </div>
-        <button class="btn " [disabled]="settingsForm.invalid " (click)="onSubmit() ">Save settings</button>
-    </form>
+  <form #settingsForm="ngForm">
+    <div *ngIf="isUserAuthenticated">
+      <mat-form-field class="example-full-width">
+        <input matInput id="username" value="User name" />
+      </mat-form-field>
+    </div>
+    <p class=" subtitle ">Buttons</p>
+    <div *ngFor="let itembtn of userBookSettings.buttonOptions">
+      <input type="checkbox" [id]="itembtn.value" [name]="itembtn.value" [value]="itembtn.value" [(ngModel)]="itembtn.checked" />
+      <label [for]="itembtn.value">
+        {{ itembtn.label }}
+        <svg viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg">
+          <path d="M 10 10 L 90 90"></path>
+          <path d="M 90 10 L 10 90"></path>
+        </svg>
+      </label>
+    </div>
+    <p class="subtitle">Info about word</p>
+    <div *ngFor="let itemword of userBookSettings.wordOptions">
+      <input type="checkbox" [id]="itemword.value" [name]="itemword.value" [value]="itemword.value" [(ngModel)]="itemword.checked" />
+      <label [for]="itemword.value">
+        {{ itemword.label }}
+        <svg viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg">
+          <path d="M 10 10 L 90 90"></path>
+          <path d="M 90 10 L 10 90"></path>
+        </svg>
+      </label>
+    </div>
+    <button class="btn" [disabled]="settingsForm.invalid" (click)="onSubmit()">Save settings</button>
+  </form>
 </section>
