+++ rslang/src/app/ebook/models/checkbox-item.model.ts
@@ -1,5 +1,11 @@
-export interface CheckboxItem {
+export class CheckboxItem {
   value: string;
   label: string;
   checked: boolean;
+
+  constructor(value: string, label: string, checked?: boolean) {
+    this.value = value;
+    this.label = label;
+    this.checked = checked || false;
+  }
 }
